Add stubs

+++ generate.js
@@ -51,7 +51,9 @@ for (const lang of active) {
         missing.push(enPage);
       }
     }
-    fs.writeFileSync(`${lang}/modules/ROOT/nav.adoc`, localNav, 'utf8');
+    if (!fs.existsSync(`${lang}/modules/ROOT/nav.adoc`)) {
+       fs.writeFileSync(`${lang}/modules/ROOT/nav.adoc`, localNav, 'utf8');
+    }
     fs.writeFileSync(`${lang}/modules/ROOT/pages/missing.adoc`, '= Missing translations\n\n * '
     + missing.join('\n * ') + '\n\n== Extra translations\n\n * '
     + orphans.join('\n * '), 'utf8');
